delete src from online_inference
