+++ online_inference/requirements/prediction_request.txt
@@ -1,0 +1,5 @@
+pandas
+requests
+yaml
+click
+marshmallow-dataclass
