+++ online_inference/requirements/testing.txt
@@ -1,0 +1,9 @@
+numpy
+pandas
+scikit-learn
+pydantic
+fastapi
+uvicorn
+pytest
+requests
+docker
