+++ online_inference/requirements/webapp.txt
@@ -1,0 +1,6 @@
+numpy
+pandas
+scikit-learn
+pydantic
+fastapi
+uvicorn
