--- online_inference/src/data/__init__.py
@@ -1,8 +1,0 @@
-from .transformations import split_dataset
-from .utils import read_dataset, read_dataset_from_json
-
-__all__ = [
-    "split_dataset",
-    "read_dataset",
-    "read_dataset_from_json",
-]
