--- online_inference/src/features/__init__.py
@@ -1,6 +1,0 @@
-from .preprocessing import build_transformer, pop_target
-
-__all__ = [
-    "build_transformer",
-    "pop_target",
-]
