--- online_inference/src/models/__init__.py
@@ -1,23 +1,0 @@
-from .evaluation import (
-    train_model,
-    predict_model,
-    evaluate_model,
-    make_predict,
-)
-from .utils import (
-    save_model_to_pickle,
-    load_model_from_pickle,
-    save_predicts_to_csv,
-    save_metrics_to_json,
-)
-
-__all__ = [
-    "train_model",
-    "evaluate_model",
-    "predict_model",
-    "save_model_to_pickle",
-    "load_model_from_pickle",
-    "save_predicts_to_csv",
-    "save_metrics_to_json",
-    "make_predict",
-]
