--- online_inference/src/parameters/__init__.py
@@ -1,21 +1,0 @@
-from .parameters import (
-    FeatureParams,
-    SplittingParams,
-    TrainingParams,
-    TrainingPipelineParams,
-    PredictionPipelineParams,
-)
-from .utils import (
-    read_training_pipeline_params,
-    read_prediction_pipeline_params,
-)
-
-__all__ = [
-    "FeatureParams",
-    "SplittingParams",
-    "TrainingParams",
-    "TrainingPipelineParams",
-    "PredictionPipelineParams",
-    "read_training_pipeline_params",
-    "read_prediction_pipeline_params",
-]
