--- online_inference/src/pipelines/__init__.py
@@ -1,7 +1,0 @@
-from .training import run_training_pipeline
-from .prediction import run_prediction_pipeline
-
-__all__ = [
-    "run_training_pipeline",
-    "run_prediction_pipeline",
-]
