--- online_inference/src/transformers/__init__.py
@@ -1,11 +1,0 @@
-from .transformers import CustomStandardScaler
-from .utils import (
-    save_transformer_to_pickle,
-    load_transformer_from_pickle,
-)
-
-__all__ = [
-    "CustomStandardScaler",
-    "save_transformer_to_pickle",
-    "load_transformer_from_pickle",
-]
