+++ online_inference/webapp/__init__.py
@@ -1,0 +1,6 @@
+from .webapp import app
+
+
+__all__ = [
+    "app",
+]
